move matplotlib printers into own module
